Adjusted supported operating systems for DP2
I removed Windows XP and Vista from the list of supported operating systems for Deluxe Pacman 2 as I have not personally tested it on them ever.  I should work on them, but there may be reasons it does not.  The game was developed and tested on Windows 7.  Tested on Windows 8 and Windows 10 and it ran without problems, so that will be the only I can verify it runs on.